feat(interface-types) Implement the `string.(lift|lower)_memory` instructions.
The `string.lift_memory` instruction replaces `memory-to-string`, and
`string.lower_memory` replaces `string-to-memory`.
